moved test.ts into lib/
